apps/questions/serializers.py
from rest_framework import serializers

from apps.questions.models import Questions


class QuestionsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Questions
        fields = "__all__"
